+++ src/test/resources/application-test.yml
@@ -15,6 +15,8 @@ spring:
   datasource:
     driver-class-name: com.mysql.cj.jdbc.Driver
   redis:
+    host: localhost
+    port: 6379
     timeout: 3000
